Refactor: common 디렉터리 파일 리팩터링
Motivation:
- #57 을 참고하세요.

Modification:
- common 디렉터리 파일 리팩터링
