+++ src/main/java/com/gamsung/backend/docs/TestController.java
@@ -1,9 +1,8 @@
-package com.gamsung.backend.global.common;
+package com.gamsung.backend.docs;
 
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
